requirements/common.txt
django==3.1.5
djangorestframework
django-cors-headers
djangorestframework-simplejwt